feat/add last feature transfer monitoring page (and add recommendation last MR)

+++ pkg/admin/gui/filter.go
@@ -3,8 +3,8 @@ package gui
 import "net/http"
 
 type Filters struct {
-	Offset           int
-	Limit            int
+	Offset           uint64
+	Limit            uint64
 	OrderAsc         bool
 	Permissions      string
 	PermissionsType  string
@@ -14,8 +14,8 @@ type Filters struct {
 }
 
 type FiltersPagination struct {
-	Offset          int
-	Limit           int
+	Offset          uint64
+	Limit           uint64
 	OrderAsc        bool
 	DisableNext     bool
 	DisablePrevious bool
@@ -25,7 +25,7 @@ type FiltersPagination struct {
 
 const DefaultLimitPagination = 30
 
-func paginationPage(filter *FiltersPagination, lenList int, r *http.Request) {
+func paginationPage(filter *FiltersPagination, lenList uint64, r *http.Request) {
 	if r.URL.Query().Get("previous") == "true" && filter.Offset > 0 {
 		filter.Offset--
 	}
